More fixes to system tests

+++ system_tests/test_cse_client.py
@@ -50,6 +50,7 @@ TODO() by priority
 """
 
 import collections
+import os
 import re
 import subprocess
 import time
@@ -92,10 +93,14 @@ def cse_server():
 
     # start cse server as subprocess
     cmd = f"cse run -c {env.ACTIVE_CONFIG_FILEPATH}"
-    p = subprocess.Popen(cmd.split(), shell=True)
-                         # stdout=subprocess.DEVNULL,
-                         # stderr=subprocess.STDOUT)
-    time.sleep(env.WAIT_INTERVAL)  # server takes a little time to set up
+    p = None
+    if os.name == 'nt':
+        p = subprocess.Popen(cmd.split(), shell=True)
+    else:
+        p = subprocess.Popen(cmd.split(),
+                             stdout=subprocess.DEVNULL,
+                             stderr=subprocess.STDOUT)
+    time.sleep(env.WAIT_INTERVAL * 3)  # server takes a little time to set up
 
     # enable kubernetes functionality on our ovdc
     # by default, an ovdc cannot deploy kubernetes clusters
@@ -132,12 +137,15 @@ def cse_server():
 
     # terminate cse server subprocess
     try:
-        # p.terminate()
-        parent_pid = p.pid
-        parent = psutil.Process(parent_pid)	
-        for child in parent.children(recursive=True):
-            child.kill()
-        parent.kill()
+        if p:
+            if os.name == 'nt':
+                parent_pid = p.pid
+                parent = psutil.Process(parent_pid)
+                for child in parent.children(recursive=True):
+                    child.kill()
+                parent.kill()
+            else:
+                p.terminate()
     except OSError:
         pass
